implemented createDocument() and createOrUpdateDocument() overloads for manually passing partition key value when needed
