--- azuredata/src/androidTest/java/com/azure/data/integration/CustomDocumentTests.kt
@@ -1,13 +1,0 @@
-package com.azure.data.integration
-
-import android.support.test.runner.AndroidJUnit4
-import com.azure.data.integration.common.CustomDocument
-import org.junit.runner.RunWith
-
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-@RunWith(AndroidJUnit4::class)
-class CustomDocumentTests : DocumentTests<CustomDocument>(CustomDocument::class.java)
